Move TaskHelper to test side only
