+++ src/Polly.Specs/Helpers/TaskHelper.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace Polly.Utilities
+namespace Polly.Specs.Helpers
 {
     /// <summary>
     /// Task helpers.
@@ -10,6 +10,6 @@ namespace Polly.Utilities
         /// <summary>
         /// Defines a completed Task for use as a completed, empty asynchronous delegate.
         /// </summary>
-        public static Task EmptyTask = Task.FromResult(true);
+        public static readonly Task EmptyTask = Task.CompletedTask; // This should now be inlined, given all targets support it. To do in its own PR, to avoid creating noise in other PRs.
     }
 }
